requirements.txt
moderngl~=5.10.0
PyGLM~=2.7.1
pygame~=2.5.2

numpy~=1.26.3

freetype-py==2.4.0

# PyWavefront~=1.3.3
